Extract common resource controller fixes to an abstract class
